+++ lib/rails_failover/active_record/railtie.rb
@@ -23,17 +23,17 @@ module RailsFailover
 
             # We are doing this manually for now since we're awaiting Rails 6.1 to be released which will
             # have more stable ActiveRecord APIs for handling multiple databases with different roles.
-            ::ActiveRecord::Base.connection_handlers[::ActiveRecord::Base.reading_role] =
+            ::ActiveRecord::Base.connection_handlers[RailsFailover::ActiveRecord.reading_role] =
               ::ActiveRecord::ConnectionAdapters::ConnectionHandler.new
 
-            ::ActiveRecord::Base.connection_handlers[::ActiveRecord::Base.writing_role].connection_pools.each do |connection_pool|
+            ::ActiveRecord::Base.connection_handlers[RailsFailover::ActiveRecord.writing_role].connection_pools.each do |connection_pool|
               if connection_pool.respond_to?(:db_config)
                 config = connection_pool.db_config.configuration_hash
               else
                 config = connection_pool.spec.config
               end
               RailsFailover::ActiveRecord.establish_reading_connection(
-                ::ActiveRecord::Base.connection_handlers[::ActiveRecord::Base.reading_role],
+                ::ActiveRecord::Base.connection_handlers[RailsFailover::ActiveRecord.reading_role],
                 config
               )
             end
@@ -43,7 +43,7 @@ module RailsFailover
             rescue ::ActiveRecord::NoDatabaseError
               # Do nothing since database hasn't been created
             rescue ::PG::Error, ::ActiveRecord::ConnectionNotEstablished
-              Handler.instance.verify_primary(::ActiveRecord::Base.writing_role)
+              Handler.instance.verify_primary(RailsFailover::ActiveRecord.writing_role)
               ::ActiveRecord::Base.connection_handler = ::ActiveRecord::Base.lookup_connection_handler(:reading)
             end
           end